requirements.txt
boto3==1.33.7
beautifulsoup4==4.12.2
celery==5.3.6
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django==4.2.8
django-cors-headers==4.3.1
django-countries==7.5.1
django-extensions==3.2.3
django-filter==23.5
django-on-heroku==1.1.2
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-database-url==2.1.0
# docx2pdf==0.1.8  # only used in a window/Mac environment
drf-spectacular==0.26.5
drf-yasg==1.21.7
google-cloud-storage==2.13.0
gunicorn==21.2.0
langchain==0.0.349
numpy==1.26.2
openai==1.3.8
pandas==2.1.3
Pillow==10.1.0
playwright==1.40.0
pyautogen==0.2.0
pymongo==4.6.1
pyspellchecker==0.7.2
python-decouple==3.8
python-dotenv==1.0.0
python-Levenshtein==0.23.0
scikit-learn==1.3.2
textblob==0.17.1
textstat==0.2
tiktoken==0.5.2